move generateValidationCodes into util file
